+++ app/src/main/AndroidManifest.xml
@@ -36,6 +36,8 @@
             </intent-filter>
         </activity>
 
+        <service android:enabled="true" android:name=".Myservice" />
+
 
         <activity android:name=".Login">
             <intent-filter>
