+++ app/src/main/res/layout/fragment_cashout.xml
@@ -16,11 +16,11 @@
             android:id="@+id/logo"
             android:layout_width="40dp"
             android:layout_height="40dp"
+            android:layout_alignParentLeft="true"
             android:layout_alignParentStart="true"
             android:layout_alignParentTop="true"
             android:layout_marginTop="17dp"
-            android:background="@drawable/app_ic"
-            android:layout_alignParentLeft="true" />
+            android:background="@drawable/app_ic" />
 
         <TextView
             android:id="@+id/c_name"
@@ -28,11 +28,11 @@
             android:layout_height="wrap_content"
             android:layout_alignTop="@+id/logo"
             android:layout_toEndOf="@+id/logo"
+            android:layout_toRightOf="@+id/logo"
             android:text="@string/title"
             android:textColor="@color/bt_nv_color"
             android:textSize="25dp"
-            android:textStyle="bold"
-            android:layout_toRightOf="@+id/logo" />
+            android:textStyle="bold" />
 
 
         <TextView
@@ -59,12 +59,12 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_alignEnd="@+id/cashout"
+            android:layout_alignRight="@+id/cashout"
             android:layout_below="@+id/cash_img"
             android:layout_marginTop="33dp"
             android:text="Available Cash : "
             android:textSize="25dp"
-            android:textStyle="bold"
-            android:layout_alignRight="@+id/cashout" />
+            android:textStyle="bold" />
 
         <TextView
             android:id="@+id/cash_amo"
@@ -74,18 +74,18 @@
             android:layout_alignBottom="@+id/ava_cash"
             android:layout_toEndOf="@+id/ava_cash"
             android:layout_toRightOf="@+id/ava_cash"
-            android:text=" ₹ 110"
+            android:text=" ₹ 0"
             android:textSize="25dp"
+            android:inputType="number"
             android:textStyle="bold" />
 
-        <Button
-            android:layout_width="180dp"
-            android:layout_height="wrap_content"
-            android:layout_below="@id/cash_amo"
-            android:layout_centerHorizontal="true"
-            android:layout_marginTop="30dp"
-            android:background="@drawable/button"
-            android:text="Transfer to Wallet" />
+
+
+
+
+
+
+
     </RelativeLayout>
